Ensure output is correct after using NIST data models.
Note: Due to a deliberate Pydantic design decision field order doesn't match
the input data, it matches the model declaration order. This is a good
decision because it enables a consistent serialization order.

However it also makes files not round-trippable, which breaks the ability to
do simple diffs between the original EDF and the precinct EDF.